Improve perf of memcpy_async on pipeline<thread>
This commit improves the performance (and probably breaks the
correctness) of pipeline<thread> in combination with memcpy_async.

The first change is absolutely necessary to recover performance:

-            if (__head == __tail) {
-                return;
-            }

This branch really destroys performance.

The second change is necessary to statistically attain the same
performance as PTX:

-                            "l"(__source + __offset)
+                            "l"(__cvta_generic_to_global(__source + __offset))

I have no clue why this makes a difference, but it (statistically) does.
Without this change, the code is slower in 50% of the cases than pure PTX.